Add search, fix hibernate type issues, fix style
Search functionality uses postgres text search, which we added a GIS index for. Currently, it just searches games by name and does not perform real-time searches. The search page can be expanded in the future by introducing more filters, for things like release date and game designer.

Hibernate was complaining about mismatching data types. The reason is that we use postgres serial4 for auto incrementing primary keys on id columns. They are 4 bytes, whereas hibernate expects id columns to use 8 bytes, so int8 or bigint. This is represented in java by long. But seeing as serial4 is just 4 bytes, we have to use integers, and instruct hibernate that the column is an integer

+++ src/main/java/com/cardboardcritic/web/RawReviewResource.java
@@ -129,7 +129,7 @@ public class RawReviewResource {
     @GET
     @Path("{id}/edit")
     @Produces(MediaType.TEXT_HTML)
-    public TemplateInstance edit(@PathParam long id) {
+    public TemplateInstance edit(@PathParam int id) {
         final RawReview rawReview = rawReviewRepo.findById(id);
         final List<Game> games = gameRepo.listAll(Sort.by("name"));
         final List<Critic> critics = criticRepo.listAll(Sort.by("name"));
@@ -143,7 +143,7 @@ public class RawReviewResource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.TEXT_HTML)
     @Transactional
-    public Response save(@PathParam long id, @BeanParam RawReviewEditForm form) {
+    public Response save(@PathParam int id, @BeanParam RawReviewEditForm form) {
         final Game game = gameRepo.createNewOrFindExisting(form.newGame, form.game);
         final Critic critic = criticRepo.createNewOrFindExisting(form.newCritic, form.critic);
         final Outlet outlet = outletRepo.createNewOrFindExisting(form.newOutlet, form.outlet);
@@ -182,7 +182,7 @@ public class RawReviewResource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.TEXT_HTML)
     @Transactional
-    public Response deny(@PathParam long id) {
+    public Response deny(@PathParam int id) {
         rawReviewRepo.update("processed = true where id = ?1", id);
         return Response.seeOther(URI.create("/raw-review")).build();
     }
@@ -192,7 +192,7 @@ public class RawReviewResource {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.TEXT_HTML)
     @Transactional
-    public Response reScrape(@PathParam long id) {
+    public Response reScrape(@PathParam int id) {
         final RawReview review = rawReviewRepo.findById(id);
         final String outlet = review.getOutlet();
 
@@ -233,7 +233,7 @@ public class RawReviewResource {
         return Response.seeOther(getEditLink(id)).build();
     }
 
-    private URI getEditLink(long id) {
+    private URI getEditLink(int id) {
         return URI.create("/raw-review/" + id + "/edit");
     }
 }
